+++ electronics_abstract_parts/AbstractDebugHeaders.py
@@ -1,0 +1,24 @@
+from .Categories import ProgrammingConnector
+from electronics_model import *
+
+
+@abstract_block
+class SwdCortexTargetConnector(ProgrammingConnector):
+  """Programming header with power and SWD (SWCLK/SWDIO/RESET) pins."""
+  def __init__(self) -> None:
+    super().__init__()
+
+    self.pwr = self.Port(VoltageSink(), [Power])  # in practice these are commonly used as sources
+    self.gnd = self.Port(Ground(), [Common])
+    self.swd = self.Port(SwdHostPort(), [Output])
+
+
+@abstract_block
+class SwdCortexTargetWithSwoTdiConnector(SwdCortexTargetConnector):
+  """SWD programming header (power + SWD) with additional generic-digital SWO and TDI pins
+  which can be used as GPIOs for side-channel data like a supplementary UART console."""
+  def __init__(self) -> None:
+    super().__init__()
+
+    self.swo = self.Port(DigitalBidir(), optional=True)
+    self.tdi = self.Port(DigitalBidir(), optional=True)
